group.tf and policies.pf: use tentant id as compartment id as required
groups.tf: Fix 400: Tenant id is not equal to compartment id

When creatipng the stack in a sub-compartment of the root compartment,
env.comparment_id is not equal to env.tenant_id, but CreateGroup
always creates groups in the tenant, not a compartment. Thus, we
must pass env.tenant_id as the compartment id to use for creating
the group. For reference, see:

https://docs.oracle.com/en-us/iaas/api/#/en/identity/20160918/Group/CreateGroup

The same is needed for policies.pf to match the tenant id of the group.